API

Index

MPITape.merge

MPITape.print_merged

MPITape.print_mytape

MPITape.readall_and_merge

References

MPITape.print_merged — Method

print_merged(tape; color)

Prints the given tape with the assumption that it contains events from multiple MPI ranks. Typically, the input should be the result of MPITape.merge(), MPITape.readall_and_merge(), or similar.

MPITape.print_mytape — Method

print_mytape(; showrank)

Print the local tape (of the calling MPI rank).

MPITape.merge — Method

To be called on all MPI ranks. Saves all tapes to disk and then reads and merges them on the master (rank 0). Returns the merged tape on the master and nothing on all other ranks.

MPITape.readall_and_merge — Function

readall_and_merge()
readall_and_merge(dir; prefix)

Reads all tapes from disk and merges them into a single "merged tape".